Extract JSON utilities into separate module
